+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Similar Changelog
 
+## 1.1.6 - 2024.04.28
+### Added
+* Added `ServicesTrait` for the plugin service component registration
+* Add `phpstan` and `ecs` code linting
+* Add `code-analysis.yaml` GitHub action
+
+### Changed
+* Updated docs to use node 20 & a new sitemap plugin
+* PHPstan code cleanup
+* ECS code cleanup
+
+### Fixed
+* Fixed an issue where an exception would be thrown if `orderBy` was `null` ([#49](https://github.com/nystudio107/craft-similar/issues/49))
+
 ## 1.1.5 - 2021.06.03
 ### Changed
 * If `orderBy` criteria is omitted, entries are now sorted by the number of relations in common. (https://github.com/nystudio107/craft-similar/issues/35)